\name{stopwoRds}
\alias{stopwoRds-package}
\alias{stopwoRds}
\docType{package}
\title{
A Selection of Stopwords in Different Languages
}
\description{
The idea behind this package is to give the user control over the stop-word 
selection.
}
\details{
The DESCRIPTION file:
The idea behind this package is to give the user control over the stop-word 
selection. The core \code{generate_stoplist} function relies on 
\code{multilingual_stopwords}, a large data frame derived from the current 
release of the Universal Dependencies Treebanks. We have included all languages 
whose corpora totalled above 10,000 tokens -- large enough to cover all common 
closed-class words, such as prepositions, conjunctions, and auxiliary verbs.
The data comes encoded in UTF-8. 
}
\author{
ABC
}
\references{
XYZ
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
