\name{stopwoRds}
\alias{stopwoRds-package}
\alias{stopwoRds}
\docType{package}
\title{
stopwoRds
}
\description{
\packageDescription{stopwoRds}

}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stopwoRds}
\packageIndices{stopwoRds}
The idea behind this package is to give the user control over the stop-word 
selection. The core \code{generate_stoplist} function relies on 
\code{multilingual_stopwords}, a large data frame derived from the current 
release of the Universal Dependencies Treebanks. We have included all languages 
whose corpora totalled above 10,000 tokens -- large enough to cover all common 
closed-class words, such as prepositions, conjunctions, and auxiliary verbs.
The data comes encoded in UTF-8. 
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
ABC
}
\references{
XYZ
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
