\name{stopwords-package}
\alias{stopwords-package}
\alias{stopwoRds}
\docType{package}
\title{
\packageTitle{stopwoRds}
}
\description{
\packageDescription{stopwords}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stopwords}
\packageIndices{stopwords}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{stopwords}

Maintainer: \packageMaintainer{stopwords}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
