\name{stopwoRds}
\alias{stopwords-package}
\alias{stopwoRds}
\docType{package}
\title{
\packageTitle{stopwoRds}
}
\description{
\packageDescription{stopwords}

}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stopwords}
\packageIndices{stopwords}
The idea behind this package is to give the user control over the stop-word 
selection. The core \code{generate_stoplist} function relies on 
\code{multilingual_stopwords}, a large data frame derived from the current 
release of the Universal Dependencies Treebanks. We have included all languages 
whose corpora totalled above 10,000 tokens -- large enough to cover all common 
closed-class words, such as prepositions, conjunctions, and auxiliary verbs.
The data comes encoded in UTF-8. 
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{stopwords}

Maintainer: \packageMaintainer{stopwords}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
